Backend-OCI/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.1
oci==2.149.0
httpx==0.27.2
pydantic==2.8.2
python-multipart==0.0.12